+++ ashen1/src/ash_str.h
@@ -12,8 +12,8 @@ char *ash_pstrdup(ash_pool_t *pool, const char *src);
 
 char *ash_pstrcat(ash_pool_t *pool, ...);
 
-size_t ash_vsnprintf(char *buf, size_t buf_sz, const char *fmt, va_list ap);
+int32_t ash_vsnprintf(char *buf, size_t buf_sz, const char *fmt, va_list ap);
 
-size_t ash_snprintf(char *buf, size_t buf_sz, const char *fmt, ...);
+int32_t ash_snprintf(char *buf, size_t buf_sz, const char *fmt, ...);
 
 #endif /* !defined(_ASH_STR_H_) */
